Created TicTacToeFrame
TicTactoeTile now extends JButton, gui is set up with a mainPanel that includes the mainLabelPanel, gameBoardPanel, and resultsPanel. The mainPanel has a border layout and the gameBoardPanel has a grid layout. Upon each click of a button an actionListener updates the board and checks to see if either player has won yet, updating the resultsPanel and gameBoardPanel as required. For some reason after the first game every subsequent game will last for one click and then the game will end with a null result.